requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.9
python-dotenv==1.1.0
SQLAlchemy==2.0.40
sqlmodel==0.0.24
asyncpg==0.30.0
Jinja2==3.1.3
aiofiles==23.2.1
python-magic==0.4.27
httpx==0.26.0
supabase==2.15.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.13.1
